requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.1.4
django-cors-headers==4.6.0
django-summernote==0.8.20.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
webencodings==0.5.1
